UPDATE: move orSeparator.tsx
