+++ packages/app/features/accounts/sign-in/signInForm.tsx
@@ -3,7 +3,7 @@ import * as Yup from "yup";
 import { SchemaOf } from "yup";
 import { ActivityIndicator, Pressable, Text, View } from "app/design-system";
 import WalletconnectBtn from "app/ui/walletconnectBtn";
-import { Separator } from "app/features/accounts/orSeparator";
+import { Separator } from "app/ui/orSeparator";
 import { Formik, FormikProps } from "formik";
 import StyledTextInput from "app/features/accounts/styledTextInput";
 import { useRequestTokenMutation } from "app/api/graphql";
